+++ src/components/common/carousel/IntroImageCarousel.tsx
