+++ src/components/domain/auth/DeleteAccountButton.tsx
@@ -1,0 +1,4 @@
+const DeleteAccountButton = () => {
+  return <button className="text-body3-r text-neutral-500">회원탈퇴</button>;
+};
+export default DeleteAccountButton;
