+++ src/components/domain/auth/index.tsx
@@ -1,2 +1,2 @@
-export { default as GoogleLoginButton } from './login/GoogleLoginButton';
-export { default as LoginSwiper } from './login/IntroImageCarousel';
+export { default as GoogleLoginButton } from './GoogleLoginButton';
+export { default as DeleteAccountButton } from './DeleteAccountButton';
