+++ src/components/domain/influencer/index.tsx
@@ -1,0 +1,1 @@
+export { default as OnePickInfluencer } from './OnePickInfluencer';
